+++ source/sass/components/article-page.scss
@@ -421,12 +421,13 @@
 }
 
 .article-navbar {
+  --margin-top: -100px;
   /* stylelint-disable */
   background: rgb(231, 231, 252); // One-off colors; no variable exists for this
   /* stylelint-enable */
   padding: 9px 0;
   position: sticky;
-  margin-top: -100px;
+  margin-top: var(--margin-top);
   left: 0;
   right: 0;
   z-index: -1;
